+++ src/main/java/eu/tango/self/adaptation/manager/model/Gpu.java
@@ -120,7 +120,7 @@ public class Gpu extends AldeJsonObjectWrapper {
      */
     public int getId() {
         if (json.has("id")) {
-            return (int) json.getInt("id");
+            return json.getInt("id");
         }
         //the default assumption is zero.
         return 0;
